NOTES 2.0 (Productive) 🔥
We are bringing a Hella lotta changes to our app. We are letting go of our previous features that carried us to improve or rather move on to new ventures and taste for the app. By deprecating old features in a phased manner, we are planning to deprecate _Notes_ to bring the updated and fit for the feature name **Productive**. The change log will be cleared to start fresh and to be the new and advanced **Productive** App